Health Bar

+++ PocketPets/Assets/CombatFiles/Pet.cs
@@ -6,29 +6,29 @@ using UnityEngine;
 public class Pet : MonoBehaviour
 {
     public int level;
-    public double maxHealth;
-    public double health;
-    public double attack;
-    public double modifier;
+    public float maxHealth;
+    public float health;
+    public float attack;
+    public float modifier;
     public double criticalChance;
     public List<string> items;
 
     private DifficultySetter difficultySetter;
-    public double enemyModifier = 1;
+    public float enemyModifier = 1;
 
 
-    public double GetAttack()
+    public float GetAttack()
     {
         DiffChecker();
         System.Random random = new System.Random();
         int playerChance = random.Next(0, 100);
         if (gameObject.CompareTag("Enemy") && criticalChance > playerChance)
         {
-            return (attack + modifier) * 1.5 * enemyModifier;
+            return (attack + modifier) * 1.5f * enemyModifier;
         }
         if (criticalChance > playerChance)
         {
-            return (attack + modifier) * 1.5;
+            return (attack + modifier) * 1.5f;
         }
         if (gameObject.CompareTag("Enemy"))
         {
@@ -43,15 +43,15 @@ public class Pet : MonoBehaviour
     {
         if (DifficultySetter.isDiffEasy == true)
         {
-            enemyModifier = 0.75;
+            enemyModifier = 0.75f;
         }
         if (DifficultySetter.isDiffNormal == true)
         {
-            enemyModifier = 1;
+            enemyModifier = 1f;
         }
         if (DifficultySetter.isDiffHard == true)
         {
-            enemyModifier = 1.25;
+            enemyModifier = 1.25f;
         }
     }
 
@@ -77,7 +77,7 @@ public class Pet : MonoBehaviour
         }
     }
 
-    public void TakeDamage(double damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
     }
